Diffusion-based-Segmentation/requirements.txt
# torch==1.9.0
# torchvision==0.10.0
# numpy==1.19.5
# tensorboard==2.5.0
Omegaconf==2.1.0
Pillow==8.2.0
scikit-learn==0.24.2
matplotlib==3.3.1
setproctitle==1.1.10
SimpleITK==2.0.1
albumentations==0.5.1
einops==0.3
pandas==1.2.2
pandasql==0.7.3
art==5.1
nibabel==3.2.1
byol_pytorch==0.5.6
visdom==0.1.8.9